Move endpoint logic to separate module
