+++ isserviceup/storage/services.py
@@ -1,20 +1,17 @@
 import time
-from typing import List
-import datetime
-import redis
 
-from isserviceup.services.models.service import Service, Status
+from isserviceup.services.models.service import Status
 
 
 _status_key = lambda service: 'service:{}'.format(service.name)
 _last_update_key = 'services:last_update'
 
 
-def set_last_update(client: redis.Redis, t: float):
+def set_last_update(client, t):
     client.set(_last_update_key, t)
 
 
-def set_service_status(client: redis.Redis, service: Service, status: Status) -> Status:
+def set_service_status(client, service, status):
     key = _status_key(service)
     pipe = client.pipeline()
     pipe.hget(key, 'status')
@@ -28,12 +25,10 @@ def set_service_status(client: redis.Redis, service: Service, status: Status) ->
     return prev_status
 
 
-def get_status(client: redis.Redis, services: List[Service]) -> (List[str], datetime.datetime):
+def get_status(client, services):
     pipe = client.pipeline()
     for service in services:
         pipe.hget(_status_key(service), 'status')
-    pipe.get(_last_update_key)
     values = pipe.execute()
-    status = [Status[x] if x else None for x in values[:-1]]
-    last_update = datetime.datetime.fromtimestamp(float(values[-1])) if values[-1] else None
-    return status, last_update
+    status = [Status[x] if x else None for x in values]
+    return status
